refactor: reorganize ta finisher task
